chinh sua db

+++ src/main/java/com/fai/ens/repository/DiaryDetailRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DiaryDetailRepository extends CrudRepository<DiaryDetail, Long> {
+public interface DiaryDetailRepository extends CrudRepository<DiaryDetail, Integer> {
 }
