+++ src/main/java/com/fai/ens/repository/DiaryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DiaryRepository extends CrudRepository<Diary, Long> {
+public interface DiaryRepository extends CrudRepository<Diary, Integer> {
 }
